* added changes based on jrose comments http://article.gmane.org/gmane.comp.lang.groovy.jsr/401
*** All uses of the new nonterminal annotation should by accompanied by
'nls!', by analogy with modifier.

> // TODO - verify that 't' is useful/correct here, used to be 'rt'
*** Yes, that's correct. ��The effect of declaratorBrackets would be a no-op
if the brackets are not present.

*** There's a lookahead (declaration)=> which should probably be
(declarationStart)=>.

*** The 'for' syntax is too back-tracky for my taste; it probably hides
ambiguities.
I don't favor supporting both {for (x in y)} and {for (x : y)}.
The former is what the Java designers would have used, if they had been
able to introduce the key word 'in'; let's do it right without worrying
about sideways compatibility.
The rule I propose is "if there are semicolons inside the for-head, we
support Java 1.4-type loops, else it's a Groovy x-in-y loop. ��And
change colon to 'in' when porting from Java."
Therefore, there's no need at all for squeezing in the new Java 5 "for"
syntax, since Groovy's is a suitable alternative.
Even if I get outvoted and we support the Java 5 syntax, I'd prefer for
it to be stuck onto the side of the grammar, later, not irretrievably
wedged now into the middle.

*** The argumentLabelStart rule has a typo: ��too many COLON tokens.

*** Also, '.' ('0'..'9')* should be '.' ('0'..'9')+ instead.